move : 보드 패키지 생성 및 관련 코드 이동
